tools/requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
RubikTwoPhase==1.1.1
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
